+++ test/manifest.js
@@ -41,3 +41,4 @@ require('./test_drop');
 require('./test_inheritance');
 require('./test_and');
 require('./test_or');
+require('./test_or_else');